Fixed issue with loading diagrams

+++ project/api/functions.py
@@ -13,7 +13,7 @@ def proccess_placement(app, user_id: str, project_id: str, track_id: str, data: 
         db.session.add(submission)
         db.session.commit()
 
-        cache_data: dict[str, dict[str, list[int]]] = {}
+        cache_data: dict[str, dict[str, list[float]]] = {}
         submissions = Submission.query.with_entities(Submission.submissions).filter_by(project_id=project_id).all() # type: ignore
 
         for _submission in submissions:
@@ -28,8 +28,8 @@ def proccess_placement(app, user_id: str, project_id: str, track_id: str, data: 
                 if not entry[0] in cache_data:
                     cache_data[entry[0]] = {"top": [], "left": []}
 
-                cache_data[entry[0]]["top"].append(pos[0])
-                cache_data[entry[0]]["left"].append(pos[1])
+                cache_data[entry[0]]["top"].append(float(pos[0]))
+                cache_data[entry[0]]["left"].append(float(pos[1]))
 
         cache_data["__track_id"] = track_id # type: ignore
         cache.set(f"answers:{project_id}", cache_data, timeout=3600)
@@ -37,7 +37,7 @@ def proccess_placement(app, user_id: str, project_id: str, track_id: str, data: 
 
 def cache_submissions(project_id: str) -> dict:
     submissions = Submission.query.with_entities(Submission.submissions).filter_by(project_id=project_id).all() # type: ignore
-    cache_data: dict[str, dict[str, list[int]]] = {}
+    cache_data: dict[str, dict[str, list[float]]] = {}
 
     for _submission in submissions:
         if not _submission[0]:
@@ -51,8 +51,8 @@ def cache_submissions(project_id: str) -> dict:
             if not entry[0] in cache_data:
                 cache_data[entry[0]] = {"top": [], "left": []}
 
-            cache_data[entry[0]]["top"].append(int(pos[0]))
-            cache_data[entry[0]]["left"].append(int(pos[1]))
+            cache_data[entry[0]]["top"].append(float(pos[0]))
+            cache_data[entry[0]]["left"].append(float(pos[1]))
 
     cache_data["__track_id"] = generate_uuid() # type: ignore
     cache.set(f"answers:{project_id}", cache_data, timeout=3600)
